[REFACTOR(MovieInfo)] Movie -> MovieInfo로 클래스 이동
