+++ app/src/main/java/woowacourse/movie/domain/MovieInfo.kt
@@ -1,6 +1,9 @@
 package woowacourse.movie.domain
 
+import android.os.Parcelable
 import androidx.annotation.DrawableRes
+import kotlinx.parcelize.Parcelize
+import java.time.LocalDate
 
 sealed class MovieInfo {
     data class Advertisement(
@@ -8,7 +11,15 @@ sealed class MovieInfo {
         val url: String,
     ) : MovieInfo()
 
-    data class MovieUnit(
-        val movie: Movie,
-    ) : MovieInfo()
+    @Parcelize
+    data class Movie(
+        val id: Long,
+        val title: String,
+        val screeningStartDate: LocalDate,
+        val screeningEndDate: LocalDate,
+        val runningTime: Int,
+        val description: String,
+        val thumbnail: Int,
+        val poster: Int,
+    ) : Parcelable, MovieInfo()
 }
